requirements.txt
# requirements.txt
Django==5.1.2
djangorestframework==3.16.0
djangorestframework-authtoken==1.2.0
openai==1.97.1
python-dotenv==1.0.0
django-cors-headers==4.7.0
# django-rest-framework-simplejwt==5.5.0
django-filter==25.1
ollama==0.5.3
requests==2.32.5
Pillow==11.3.0